+++ json_annotation/lib/src/allowed_keys_helpers.dart
@@ -19,7 +19,9 @@ void $checkAllowedKeys(Map map, Iterable<String> allowedKeys) {
 ///
 /// Should not be used directly.
 void $checkKeys(Map map,
-    {List<String> allowedKeys, List<String> requiredKeys}) {
+    {List<String> allowedKeys,
+    List<String> requiredKeys,
+    List<String> disallowNullValues}) {
   if (map != null && allowedKeys != null) {
     var invalidKeys =
         map.keys.cast<String>().where((k) => !allowedKeys.contains(k)).toList();
@@ -32,6 +34,18 @@ void $checkKeys(Map map,
     var missingKeys = requiredKeys.where((k) => !map.keys.contains(k)).toList();
     if (missingKeys.isNotEmpty) {
       throw new MissingRequiredKeysException(missingKeys, map);
+    }
+  }
+
+  if (map != null && disallowNullValues != null) {
+    var nullValuedKeys = map.entries
+        .where((entry) =>
+            disallowNullValues.contains(entry.key) && entry.value == null)
+        .map((entry) => entry.key as String)
+        .toList();
+
+    if (nullValuedKeys.isNotEmpty) {
+      throw new DisallowedNullValueException(nullValuedKeys, map);
     }
   }
 }
@@ -77,4 +91,16 @@ class MissingRequiredKeysException extends BadKeyException {
   MissingRequiredKeysException(this.missingKeys, Map map)
       : assert(missingKeys.isNotEmpty),
         super._(map);
+}
+
+/// Exception thrown if there are keys with disallowed `null` values in a JSON
+/// map that was provided during deserialization.
+class DisallowedNullValueException extends BadKeyException {
+  final List<String> keysWithNullValues;
+
+  DisallowedNullValueException(this.keysWithNullValues, Map map) : super._(map);
+
+  @override
+  String get message => 'These keys had `null` values, '
+      'which is not allowed: $keysWithNullValues';
 }