+++ json_serializable/test/test_files/json_test_example.dart
@@ -42,6 +42,8 @@ enum Category { top, bottom, strange, charmed, up, down }
 
 @JsonSerializable()
 class Order extends Object with _$OrderSerializerMixin {
+  /// Used to test that `disallowNullValues: true` forces `includeIfNull: false`
+  @JsonKey(disallowNullValue: true)
   int count;
   bool isRushed;