+++ json_serializable/test/test_files/json_test_example.non_nullable.g.dart
@@ -41,6 +41,7 @@ abstract class _$PersonSerializerMixin {
 }
 
 Order _$OrderFromJson(Map<String, dynamic> json) {
+  $checkKeys(json, disallowNullValues: const ['count']);
   return new Order(
       $enumDecode('Category', Category.values, json['category'] as String),
       (json['items'] as List)